+++ packages/redux-store/src/slices/connections/index.ts
@@ -1,4 +1,3 @@
 export { connectionsSlice } from './connectionsSlice'
 export { ConnectionThunks } from './connectionsThunks'
 export { ConnectionsSelectors } from './connectionsSelectors'
-export { startConnectionsListener } from './connectionsListener'